Add files via upload
Initial data collection was conducted through and API call. Data was the filtered for optimal features and then passed into a new data frame. This data frame will later be combined with the data collected through webscrapping.  In this particular file, some initial wrangling has been conducted to replace any missing data within the payload with the average value of the payloads.