Add files via upload
This file is labeled part 5 as it was the 5th item to appear in the project. For some reason, this appears before part 4 in the PowerPoint requirements. I am unsure as to the reason for this. Regardless, in this section, basic visualization is conducted. Using MatplotLib and Seaborne, we look at the data and compare various features with the outcome of launches.