Add files via upload
Similarly to part 5, this section appeared in the course as part 4, but appeared in the PowerPoint instructions as part 5. Regardless, SQL queries were performed in this section to extract information useful for narrowing down where to find relevant features.

NOTE: This is the  re-upload of the original file. I failed to notice some parts of the questions and thus had incorrect answers. I have rectified this issue.